# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Schuba API'
  description: 'API for managing diving club activities'
  contact:
    name: 'Groupe 1'
  version: 0.9.0
paths:
...
  /api/club:
    get:
      tags:
        - Clubs
      summary: 'Get a list of clubs'
      description: 'Fetches clubs based on the provided filters (ID or Name).'
      operationId: getClubs
      parameters:
        -
          name: ID
          in: query
          description: 'The ID of the club'
          required: false
          schema:
            type: integer
            format: int64
        -
          name: NAME
          in: query
          description: 'Partial name of the club to search for'
          required: false
          schema:
            type: string
            example: 'Sports Club'
      responses:
        200:
          description: 'List of clubs successfully fetched'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    CLU_ID:
                      type: integer
                      example: 1
                    CLU_NOM:
                      type: string
                      example: 'Sports Club'
                  type: object
        400:
          description: 'Bad request, invalid parameters'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Invalid request parameters'
                type: object
    post:
      tags:
        - Clubs
      summary: 'Create a new club'
      description: 'Creates a new club with the specified name.'
      operationId: createClub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - NAME
              properties:
                NAME:
                  description: 'The name of the club'
                  type: string
                  example: 'New Club'
              type: object
      responses:
        201:
          description: 'Club successfully created'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Club successfully created!'
                  club:
                    properties:
                      CLU_ID:
                        type: integer
                        example: 1
                      CLU_NOM:
                        type: string
                        example: 'New Club'
                    type: object
                type: object
        400:
          description: 'Invalid input'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The given data was invalid.'
                type: object
  '/api/club/{id}':
    put:
      tags:
        - Clubs
      summary: 'Update an existing club'
      description: "Updates a club's name by its ID."
      operationId: updateClub
      parameters:
        -
          name: id
          in: path
          description: 'ID of the club to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - NAME
              properties:
                NAME:
                  description: 'The name of the club'
                  type: string
                  example: 'New Club Name'
              type: object
      responses:
        200:
          description: 'Club successfully updated'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Club successfully updated!'
                  club:
                    properties:
                      CLU_ID:
                        type: integer
                        example: 1
                      CLU_NOM:
                        type: string
                        example: 'New Club Name'
                    type: object
                type: object
        404:
          description: 'Club not found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Club not found.'
                type: object
    delete:
      tags:
        - Clubs
      summary: 'Delete a club'
      description: 'Deletes a club by its ID.'
      operationId: deleteClub
      parameters:
        -
          name: id
          in: path
          description: 'ID of the club to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: 'Club successfully deleted'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Club successfully deleted!'
                type: object
        404:
          description: 'Club not found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Club not found.'
                type: object
...